refactor: move user
